requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.5.1
scikit-learn~=1.2.0
pyyaml~=5.4.1
setuptools~=65.6.3